YARP_math: Remove the possibility to build with gsl
