+++ .travis.yml
@@ -258,9 +258,9 @@ install:
   - if [ "$TRAVIS_OS_NAME" == "osx" ]; then brew install qt5; fi
   - if [ "$TRAVIS_OS_NAME" == "osx" ]; then export Qt5_DIR="/usr/local/opt/qt5/lib/cmake/Qt5"; fi
 
-  # Install GPL dependencies (gsl, qcustomplot)
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y libgsl0-dev libqcustomplot-dev; fi
-  - if [ "$TRAVIS_OS_NAME" == "osx" ]; then brew install gsl readline; fi
+  # Install GPL dependencies (qcustomplot)
+  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y libqcustomplot-dev; fi
+  - if [ "$TRAVIS_OS_NAME" == "osx" ]; then brew install readline; fi
 
   # Install other dependencies (opencv, eigen)
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then sudo apt-get install -y libopencv-dev libeigen3-dev; fi
@@ -294,7 +294,6 @@ before_script:
   - if $TRAVIS_WITH_VALGRIND; then export YARP_CMAKE_OPTIONS="${YARP_CMAKE_OPTIONS} -DYARP_VALGRIND_TESTS=ON -DYARP_VALGRIND_DRD_TESTS=OFF -DYARP_VALGRIND_HELGRIND_TESTS=OFF YARP_VALGRIND_MEMCHECK_TESTS=ON YARP_VALGRIND_SGCHECK_TESTS=OFF"; fi
   - if $TRAVIS_NO_DEPRECATED; then export YARP_CMAKE_OPTIONS="${YARP_CMAKE_OPTIONS} -DYARP_NO_DEPRECATED=ON"; fi
   - if $TRAVIS_WITH_BROKEN_TESTS; then export YARP_CMAKE_OPTIONS="${YARP_CMAKE_OPTIONS} -DYARP_ENABLE_BROKEN_TESTS=ON"; fi
-  - export YARP_CMAKE_OPTIONS="${YARP_CMAKE_OPTIONS} -DCREATE_LIB_MATH_USING_GSL=OFF"
 
   # Print cmake flags
   - echo "--------------------     BEGIN CMAKE FLAGS    --------------------"
