+++ cmake/YarpFindDependencies.cmake
@@ -252,7 +252,6 @@ mark_as_advanced(SKIP_ACE)
 
 
 option(CREATE_LIB_MATH "Create math library libYARP_math?" OFF)
-cmake_dependent_option(CREATE_LIB_MATH_USING_GSL "Create math library libYARP_math using GSL?" OFF CREATE_LIB_MATH ON)
 cmake_dependent_option(CREATE_YARPROBOTINTERFACE "Do you want to compile yarprobotinterface?" ON YARP_COMPILE_EXECUTABLES OFF)
 cmake_dependent_option(CREATE_YARPMANAGER_CONSOLE "Do you want to compile YARP Module Manager (console)?" ON YARP_COMPILE_EXECUTABLES OFF)
 cmake_dependent_option(CREATE_YARPDATADUMPER "Do you want to compile yarpdatadumper?" ON YARP_COMPILE_EXECUTABLES OFF)
@@ -271,6 +270,7 @@ yarp_renamed_option(CREATE_GYARPMANAGER CREATE_YARPMANAGER)
 yarp_deprecated_option(CREATE_GYARPBUILDER)
 yarp_deprecated_option(CREATE_YARPBUILDER)
 yarp_deprecated_option(CREATE_YARPMANAGER_PP) # Since YARP 2.3.68
+yarp_deprecated_option(CREATE_LIB_MATH_USING_GSL) # Since YARP 2.3.70
 
 if(CREATE_YARPMANAGER_CONSOLE OR CREATE_YARPMANAGER)
   set(CREATE_LIB_MANAGER ON CACHE INTERNAL "Create manager library libYARP_manager?")
@@ -327,13 +327,8 @@ find_package(Libedit QUIET)
 checkandset_dependency(Libedit)
 
 if(CREATE_LIB_MATH)
-  if(CREATE_LIB_MATH_USING_GSL)
-    find_package(GSL QUIET)
-    checkandset_dependency(GSL)
-  else()
-    find_package(Eigen3 QUIET)
-    checkandset_dependency(Eigen3)
-  endif()
+  find_package(Eigen3 QUIET)
+  checkandset_dependency(Eigen3)
 endif()
 
 if(CREATE_YARPROBOTINTERFACE OR CREATE_YARPSCOPE OR CREATE_LIB_MANAGER)
@@ -421,7 +416,6 @@ print_dependency(YCM)
 print_dependency(ACE)
 print_dependency(SQLite)
 print_dependency(Eigen3)
-print_dependency(GSL)
 print_dependency(TinyXML)
 print_dependency(xmlrpcpp)
 print_dependency(Qt5)
@@ -451,11 +445,7 @@ print_dependency(Stage)
 
 check_skip_dependency(SKIP_ACE ACE)
 check_required_dependency(SQLite)
-check_alternative_dependency(CREATE_LIB_MATH Eigen3 GSL)
-if(CREATE_LIB_MATH)
-  check_optional_dependency(CREATE_LIB_MATH_USING_GSL GSL)
-  check_skip_dependency(CREATE_LIB_MATH_USING_GSL Eigen3)
-endif()
+check_optional_dependency(CREATE_LIB_MATH Eigen3)
 check_optional_dependency(CREATE_LIB_MANAGER TinyXML)
 check_optional_dependency(CREATE_YARPSCOPE TinyXML)
 check_optional_dependency(CREATE_GUIS Qt5)
