+++ src/libYARP_eigen/CMakeLists.txt
@@ -12,4 +12,9 @@ set_property(GLOBAL APPEND PROPERTY YARP_TREE_INCLUDE_DIRS "${PROJECT_SOURCE_DIR
 
 # When we support CMake 3.0 as the minimum required, we can make this an INTERFACE library
 # For now we just install the headers in a specific location
-install(DIRECTORY ${PROJECT_SOURCE_DIR}/include/yarp/eigen DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/yarp)
+if(NOT CMAKE_MINIMUM_REQUIRED_VERSION VERSION_LESS 3.0)
+  message(AUTHOR_WARNING "CMAKE_MINIMUM_REQUIRED_VERSION is now ${CMAKE_MINIMUM_REQUIRED_VERSION}. YARP_eigen can now be an INTERFACE library.")
+endif()
+
+install(DIRECTORY ${PROJECT_SOURCE_DIR}/include/yarp/eigen
+        DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/yarp)
