+++ src/libYARP_math/include/yarp/math/Math.h
@@ -16,9 +16,9 @@
 namespace yarp
 {
      /**
-     * Mathematical operations, mostly a GSL wrapper.
+     * Mathematical operations.
      */
-    namespace math 
+    namespace math
     {
         /**
         * Addition operator between a scalar and a vector (defined in Math.h).
@@ -56,8 +56,8 @@ namespace yarp
         /**
         * Addition operator between matrices, returns a+b (defined in Math.h).
         */
-        YARP_math_API yarp::sig::Matrix& operator+=(yarp::sig::Matrix &a, const yarp::sig::Matrix &b);        
-        
+        YARP_math_API yarp::sig::Matrix& operator+=(yarp::sig::Matrix &a, const yarp::sig::Matrix &b);
+
         /**
         * Subtraction operator between a vector and a scalar (defined in Math.h).
         * Subtract the scalar to all the elements of the vector.
@@ -94,10 +94,10 @@ namespace yarp
         /**
         * Subtraction operator between matrices, returns a-b (defined in Math.h).
         */
-        YARP_math_API yarp::sig::Matrix& operator-=(yarp::sig::Matrix &a, const yarp::sig::Matrix &b);                
-        
+        YARP_math_API yarp::sig::Matrix& operator-=(yarp::sig::Matrix &a, const yarp::sig::Matrix &b);
+
         /**
-        * Scalar-vector product operator (defined in Math.h). 
+        * Scalar-vector product operator (defined in Math.h).
         * @param k a scalar
         * @param b vector
         * @return k*b
@@ -105,15 +105,15 @@ namespace yarp
         YARP_math_API yarp::sig::Vector operator*(double k, const yarp::sig::Vector &b);
 
         /**
-        * Vector-scalar product operator (defined in Math.h). 
+        * Vector-scalar product operator (defined in Math.h).
         * @param b a vector
         * @param k a scalar
         * @return b*k
         */
         YARP_math_API yarp::sig::Vector operator*(const yarp::sig::Vector &b, double k);
-        
+
         /**
-        * Vector-scalar product operator (defined in Math.h). 
+        * Vector-scalar product operator (defined in Math.h).
         * @param b a vector
         * @param k a scalar
         * @return b*k
@@ -121,33 +121,33 @@ namespace yarp
         YARP_math_API yarp::sig::Vector& operator*=(yarp::sig::Vector &b, double k);
 
         /**
-        * Vector-matrix product operator (defined in Math.h). 
+        * Vector-matrix product operator (defined in Math.h).
         * @param a is a vector (interpreted as a row)
         * @param m is a matrix
         * @return a*m
         */
-        YARP_math_API yarp::sig::Vector operator*(const yarp::sig::Vector &a, 
+        YARP_math_API yarp::sig::Vector operator*(const yarp::sig::Vector &a,
             const yarp::sig::Matrix &m);
 
         /**
-        * Vector-matrix product operator (defined in Math.h). 
+        * Vector-matrix product operator (defined in Math.h).
         * @param a is a vector (interpreted as a row)
         * @param m is a matrix
         * @return a*m
         */
         YARP_math_API yarp::sig::Vector& operator*=(yarp::sig::Vector &a, const yarp::sig::Matrix &m);
-        
+
         /**
-        * Matrix-vector product operator (defined in Math.h). 
+        * Matrix-vector product operator (defined in Math.h).
         * @param a is a vector (interpreted as a column)
         * @param m is a matrix
         * @return m*a
         */
-         YARP_math_API yarp::sig::Vector operator*(const yarp::sig::Matrix &m, 
+         YARP_math_API yarp::sig::Vector operator*(const yarp::sig::Matrix &m,
             const yarp::sig::Vector &a);
-        
+
         /**
-        * Matrix-matrix product operator (defined in Math.h). 
+        * Matrix-matrix product operator (defined in Math.h).
         * @param a a matrix
         * @param b a matrix
         * @return a*b
@@ -155,7 +155,7 @@ namespace yarp
         YARP_math_API yarp::sig::Matrix operator*(const yarp::sig::Matrix &a, const yarp::sig::Matrix &b);
 
         /**
-        * Matrix-matrix product operator (defined in Math.h). 
+        * Matrix-matrix product operator (defined in Math.h).
         * @param a a matrix
         * @param b a matrix
         * @return a*b
@@ -163,15 +163,15 @@ namespace yarp
         YARP_math_API yarp::sig::Matrix& operator*=(yarp::sig::Matrix &a, const yarp::sig::Matrix &b);
 
         /**
-       * Scalar-matrix product operator (defined in Math.h). 
+       * Scalar-matrix product operator (defined in Math.h).
        * @param k a scalar
        * @param M a matrix
        * @return k*M
        */
         YARP_math_API yarp::sig::Matrix operator*(const double k, const yarp::sig::Matrix &M);
-               
+
         /**
-       * Matrix-scalar product operator (defined in Math.h). 
+       * Matrix-scalar product operator (defined in Math.h).
        * @param M a matrix
        * @param k a scalar
        * @return M*k
@@ -179,15 +179,15 @@ namespace yarp
         YARP_math_API yarp::sig::Matrix operator*(const yarp::sig::Matrix &M, const double k);
 
         /**
-        * Matrix-scalar product operator (defined in Math.h). 
+        * Matrix-scalar product operator (defined in Math.h).
         * @param M a matrix
         * @param k a scalar
         * @return M*k
         */
         YARP_math_API yarp::sig::Matrix& operator*=(yarp::sig::Matrix &M, const double k);
-        
+
         /**
-       * Vector-vector element-wise product operator (defined in Math.h). 
+       * Vector-vector element-wise product operator (defined in Math.h).
        * @param a a vector
        * @param b a vector
        * @return a.*b (matlab notation)
@@ -195,15 +195,15 @@ namespace yarp
         YARP_math_API yarp::sig::Vector operator*(const yarp::sig::Vector &a, const yarp::sig::Vector &b);
 
         /**
-       * Vector-vector element-wise product operator (defined in Math.h). 
+       * Vector-vector element-wise product operator (defined in Math.h).
        * @param a a vector
        * @param b a vector
        * @return a.*b (matlab notation)
        */
         YARP_math_API yarp::sig::Vector& operator*=(yarp::sig::Vector &a, const yarp::sig::Vector &b);
-                
+
         /**
-       * Vector-vector element-wise division operator (defined in Math.h). 
+       * Vector-vector element-wise division operator (defined in Math.h).
        * @param a a vector
        * @param b a vector
        * @return a./b (matlab notation)
@@ -211,7 +211,7 @@ namespace yarp
         YARP_math_API yarp::sig::Vector operator/(const yarp::sig::Vector &a, const yarp::sig::Vector &b);
 
         /**
-       * Vector-vector element-wise division operator (defined in Math.h). 
+       * Vector-vector element-wise division operator (defined in Math.h).
        * @param a a vector
        * @param b a vector
        * @return a./b (matlab notation)
@@ -219,7 +219,7 @@ namespace yarp
         YARP_math_API yarp::sig::Vector& operator/=(yarp::sig::Vector &a, const yarp::sig::Vector &b);
 
         /**
-        * Vector-scalar division operator (defined in Math.h). 
+        * Vector-scalar division operator (defined in Math.h).
         * @param b a vector
         * @param k a scalar
         * @return b/k
@@ -227,7 +227,7 @@ namespace yarp
         YARP_math_API yarp::sig::Vector operator/(const yarp::sig::Vector &b, double k);
 
         /**
-        * Vector-scalar division operator (defined in Math.h). 
+        * Vector-scalar division operator (defined in Math.h).
         * @param b a vector
         * @param k a scalar
         * @return b/k
@@ -235,7 +235,7 @@ namespace yarp
         YARP_math_API yarp::sig::Vector& operator/=(yarp::sig::Vector &b, double k);
 
         /**
-        * Matrix-scalar division operator (defined in Math.h). 
+        * Matrix-scalar division operator (defined in Math.h).
         * @param M a matrix
         * @param k a scalar
         * @return M./k (matlab notation)
@@ -243,7 +243,7 @@ namespace yarp
         YARP_math_API yarp::sig::Matrix operator/(const yarp::sig::Matrix &M, double k);
 
         /**
-        * Matrix-scalar division operator (defined in Math.h). 
+        * Matrix-scalar division operator (defined in Math.h).
         * @param M a matrix
         * @param k a scalar
         * @return M./k (matlab notation)
@@ -264,7 +264,7 @@ namespace yarp
          * @param v a vector m
          * @return a matrix (n+1)Xm
          */
-        YARP_math_API yarp::sig::Matrix pile(const yarp::sig::Vector &v, const yarp::sig::Matrix &m);        
+        YARP_math_API yarp::sig::Matrix pile(const yarp::sig::Vector &v, const yarp::sig::Matrix &m);
 
         /**
          * Matrix-Vector concatenation (defined in Math.h). Add a row at the end of a matrix.
@@ -296,7 +296,7 @@ namespace yarp
          * @param v a vector n
          * @return a matrix nX(m+1)
          */
-        YARP_math_API yarp::sig::Matrix cat(const yarp::sig::Matrix &m, const yarp::sig::Vector &v);        
+        YARP_math_API yarp::sig::Matrix cat(const yarp::sig::Matrix &m, const yarp::sig::Vector &v);
 
         /**
          * Vector-Matrix concatenation (defined in Math.h). Add a column at the beginning of a matrix.
@@ -343,9 +343,9 @@ namespace yarp
         YARP_math_API yarp::sig::Vector cat(double s1, double s2, double s3, double s4);
 
         YARP_math_API yarp::sig::Vector cat(double s1, double s2, double s3, double s4, double s5);
-        
+
         /**
-        * Scalar product between vectors (defined in Math.h). 
+        * Scalar product between vectors (defined in Math.h).
         * @return a^T*b, where a and b are column vectors
         */
         YARP_math_API double dot(const yarp::sig::Vector &a, const yarp::sig::Vector &b);
@@ -380,30 +380,30 @@ namespace yarp
         YARP_math_API bool crossProductMatrix(const yarp::sig::Vector &v, yarp::sig::Matrix &res);
 
         /**
-        * Returns the Euclidean norm of the vector (defined in Math.h). 
-        * @param v is the input vector. 
-        * @return ||v||. 
+        * Returns the Euclidean norm of the vector (defined in Math.h).
+        * @param v is the input vector.
+        * @return ||v||.
         */
         YARP_math_API double norm(const yarp::sig::Vector &v);
 
         /**
-        * Returns the Euclidean squared norm of the vector (defined in Math.h). 
-        * @param v is the input vector. 
-        * @return ||v||^2. 
+        * Returns the Euclidean squared norm of the vector (defined in Math.h).
+        * @param v is the input vector.
+        * @return ||v||^2.
         */
         YARP_math_API double norm2(const yarp::sig::Vector &v);
 
         /**
         * Returns the maximum of the elements of a real vector (defined in Math.h).
-        * @param v is the input vector. 
-        * @return max(v). 
+        * @param v is the input vector.
+        * @return max(v).
         */
         YARP_math_API double findMax(const yarp::sig::Vector &v);
 
         /**
         * Returns the minimum of the elements of a real vector (defined in Math.h).
-        * @param v is the input vector. 
-        * @return min(v). 
+        * @param v is the input vector.
+        * @return min(v).
         */
         YARP_math_API double findMin(const yarp::sig::Vector &v);
 
@@ -465,40 +465,33 @@ namespace yarp
         */
         YARP_math_API bool eigenValues(const yarp::sig::Matrix& in, yarp::sig::Vector &real, yarp::sig::Vector &img);
 
-#ifndef YARP_NO_DEPRECATED
         /**
-        * @deprecated since YARP 2.3.65
-        */
-        YARP_math_DEPRECATED_API bool eingenValues(const yarp::sig::Matrix& in, yarp::sig::Vector &real, yarp::sig::Vector &img);
-#endif // YARP_NO_DEPRECATED
-
-        /**
-        * Invert a symmetric and positive definite matrix using Cholesky decomposition (defined in Math.h). 
+        * Invert a symmetric and positive definite matrix using Cholesky decomposition (defined in Math.h).
         * @param in symmetric and positive definite matrix
         * @return the inverse of the matrix
         */
         //YARP_math_API yarp::sig::Matrix chinv(const yarp::sig::Matrix& in);
 
         /**
-        * Returns the sign of a real number: 1 if positive, 
+        * Returns the sign of a real number: 1 if positive,
         * -1 if negative, 0 if equal to zero (defined in Math.h)
-        * @param v is a real number. 
-        * @return sign(v). 
+        * @param v is a real number.
+        * @return sign(v).
         */
         YARP_math_API double sign(const double &v);
 
         /**
-        * Returns the sign vector of a real vector, that is a vector 
-        * with 1 if the value is positive, 
-        * -1 if negative, 0 if equal to zero (defined in Math.h). 
-        * @param v is the input vector. 
-        * @return sign(v). 
+        * Returns the sign vector of a real vector, that is a vector
+        * with 1 if the value is positive,
+        * -1 if negative, 0 if equal to zero (defined in Math.h).
+        * @param v is the input vector.
+        * @return sign(v).
         */
         YARP_math_API yarp::sig::Vector sign(const yarp::sig::Vector &v);
 
         /**
-        * Converts a dcm (direction cosine matrix) rotation matrix R to 
-        * axis/angle representation (defined in Math.h). 
+        * Converts a dcm (direction cosine matrix) rotation matrix R to
+        * axis/angle representation (defined in Math.h).
         * @param R is the input matrix.
         * @return 4 by 1 vector for the axis/angle representation.
         */
@@ -514,9 +507,9 @@ namespace yarp
         YARP_math_API yarp::sig::Matrix axis2dcm(const yarp::sig::Vector &v);
 
         /**
-        * Converts a dcm (direction cosine matrix) rotation matrix to 
+        * Converts a dcm (direction cosine matrix) rotation matrix to
         * euler angles (ZYZ) (defined in Math.h). Three angles are returned in a vector with
-        * the following format: 
+        * the following format:
         * \f[  \mathbf{v} = [\alpha, \beta, \gamma ]\f]
         * such that the returned matrix satisfies the following:
         * \f[  R = R_z(\alpha) R_y(\beta) R_z(\gamma) \f]
@@ -526,24 +519,24 @@ namespace yarp
         YARP_math_API yarp::sig::Vector dcm2euler(const yarp::sig::Matrix &R);
 
         /**
-        * Converts euler angles (ZYZ) vector in the corresponding dcm 
-        * (direction cosine matrix) rotation matrix (defined in Math.h). The three euler 
+        * Converts euler angles (ZYZ) vector in the corresponding dcm
+        * (direction cosine matrix) rotation matrix (defined in Math.h). The three euler
         * angles are specified in a vector with the following structure:
         * \f[  \mathbf{v} = [\alpha, \beta, \gamma ]\f]
         * and the returned matrix is:
         * \f[  R = R_z(\alpha) R_y(\beta) R_z(\gamma) \f]
-        * @param euler is the input vector (alpha=z-rotation, 
+        * @param euler is the input vector (alpha=z-rotation,
         *              beta=y-rotation, gamma=z-rotation).
-        * @return 4 by 4 homogeneous matrix representing the ZYZ 
+        * @return 4 by 4 homogeneous matrix representing the ZYZ
         *         rotation with the rotation components in the top left
         *         3 by 3 submatrix.
         */
         YARP_math_API yarp::sig::Matrix euler2dcm(const yarp::sig::Vector &euler);
 
         /**
-        * Converts a dcm (direction cosine matrix) rotation matrix to 
-        * roll-pitch-yaw angles (defined in Math.h). Three angles are returned in a vector 
-        * with the following format: 
+        * Converts a dcm (direction cosine matrix) rotation matrix to
+        * roll-pitch-yaw angles (defined in Math.h). Three angles are returned in a vector
+        * with the following format:
         * \f[  \mathbf{v} = [\psi, \theta, \phi ]\f]
         * such that the returned matrix satisfies the following:
         * \f[  R = R_z(\phi) R_y(\theta) R_x(\psi) \f]
@@ -553,50 +546,49 @@ namespace yarp
         YARP_math_API yarp::sig::Vector dcm2rpy(const yarp::sig::Matrix &R);
 
         /**
-        * Converts roll-pitch-yaw angles in the corresponding dcm 
-        * (direction cosine matrix) rotation matrix (defined in Math.h). The three angles 
-        * are specified in a vector with the following structure: 
+        * Converts roll-pitch-yaw angles in the corresponding dcm
+        * (direction cosine matrix) rotation matrix (defined in Math.h). The three angles
+        * are specified in a vector with the following structure:
         * \f[  \mathbf{v} = [\psi, \theta, \phi ]\f]
         * and the returned matrix is:
         * \f[  R = R_z(\phi) R_y(\theta) R_x(\psi) \f]
-        * @param rpy is the input vector (\psi=roll 
+        * @param rpy is the input vector (\psi=roll
         *            x-rotation,\theta=pitch y-rotation, \phi=yaw
         *            z-rotation).
-        * @return 4 by 4 homogeneous matrix representing the ZYX 
+        * @return 4 by 4 homogeneous matrix representing the ZYX
         *         rotation with the rotation components in the top left
-        *         3 by 3 submatrix. 
+        *         3 by 3 submatrix.
         */
         YARP_math_API yarp::sig::Matrix rpy2dcm(const yarp::sig::Vector &rpy);
 
         /**
-        * Returns the inverse of a 4 by 4 rototranslational matrix (defined in Math.h). 
+        * Returns the inverse of a 4 by 4 rototranslational matrix (defined in Math.h).
         * @param H is the 4 by 4 rototranslational matrix.
-        * @return inverse of 4 by 4 rototranslational matrix. 
-        *  
-        * @note about 5 times faster than pinv() 
+        * @return inverse of 4 by 4 rototranslational matrix.
+        *
+        * @note about 5 times faster than pinv()
         */
         YARP_math_API yarp::sig::Matrix SE3inv(const yarp::sig::Matrix &H);
 
         /**
-        * Returns the adjoint matrix of a given roto-translational 
-        * matrix (defined in Math.h). The adjoint is a (6x6) matrix: [R , S(r)*R; 0, R] 
+        * Returns the adjoint matrix of a given roto-translational
+        * matrix (defined in Math.h). The adjoint is a (6x6) matrix: [R , S(r)*R; 0, R]
         * where R is the rotational part of H, and r the translational part.
         * @param H is the 4 by 4 rototranslational matrix.
-        * @return the adjoint matrix 
+        * @return the adjoint matrix
         */
         YARP_math_API yarp::sig::Matrix adjoint(const yarp::sig::Matrix &H);
 
         /**
-        * Returns the inverse of the adjoint matrix of a given 
-        * roto-translational matrix (defined in Math.h). The inverse of an adjoint is a 
-        * (6x6) matrix: [R^T , -S(R^T*r)*R^T; 0 , R^T] where R is the 
-        * rotational part of H, and r the translational part. 
+        * Returns the inverse of the adjoint matrix of a given
+        * roto-translational matrix (defined in Math.h). The inverse of an adjoint is a
+        * (6x6) matrix: [R^T , -S(R^T*r)*R^T; 0 , R^T] where R is the
+        * rotational part of H, and r the translational part.
         * @param H is the 4 by 4 rototranslational matrix.
-        * @return the inverse of the adjoint matrix  
+        * @return the inverse of the adjoint matrix
         */
         YARP_math_API yarp::sig::Matrix adjointInv(const yarp::sig::Matrix &H);
     }
 }
 
 #endif
-
