+++ src/libYARP_math/include/yarp/math/SVD.h
@@ -14,20 +14,20 @@
 
 namespace yarp
 {
-    namespace math 
+    namespace math
     {
-        /** 
+        /**
         * Factorize the M-by-N matrix 'in' into the singular value decomposition in = U S V^T (defined in SVD.h).
         * The diagonal elements of the singular value matrix S are stored in the vector S.
-        * The singular values are non-negative and form a non-increasing sequence from S_1 to S_N. 
-        * The matrix V contains the elements of V in untransposed form. To form the product U S V^T it 
+        * The singular values are non-negative and form a non-increasing sequence from S_1 to S_N.
+        * The matrix V contains the elements of V in untransposed form. To form the product U S V^T it
         * is necessary to take the transpose of V.
         * Defining K as min(M, N) the the input matrices are:
         * @param in input M-by-N matrix to decompose
         * @param U output M-by-K orthogonal matrix
         * @param S output K-dimensional vector containing the diagonal entries of the diagonal matrix S
         * @param V output N-by-K orthogonal matrix
-        * @note The routine computes the \a thin version of the SVD. Mathematically, the \a full SVD is 
+        * @note The routine computes the \a thin version of the SVD. Mathematically, the \a full SVD is
         *       defined with U and V as square orthogonal matrices and S as an M-by-N diagonal matrix.
         *       If U, S, V do not have the expected sizes they are resized automatically.
         *
@@ -35,7 +35,7 @@ namespace yarp
         *       function uses the Golub-Reinsch SVD algorithm.
         *       When compiled with Eigen backend, this function uses the Jacobi SVD algorithm.
         */
-        void YARP_math_API SVD(const yarp::sig::Matrix &in, 
+        void YARP_math_API SVD(const yarp::sig::Matrix &in,
             yarp::sig::Matrix &U,
             yarp::sig::Vector &S,
             yarp::sig::Matrix &V);
@@ -47,7 +47,7 @@ namespace yarp
         *       function uses the Modified Golub-Reinsch SVD algorithm (fast for M>>N)
         *       When compiled with Eigen backend, this function uses the Jacobi SVD algorithm.
         */
-        void YARP_math_API SVDMod(const yarp::sig::Matrix &in, 
+        void YARP_math_API SVDMod(const yarp::sig::Matrix &in,
             yarp::sig::Matrix &U,
             yarp::sig::Vector &S,
             yarp::sig::Matrix &V);
@@ -58,14 +58,14 @@ namespace yarp
         *
         * @note For both Eigen and GSL backends, the Jacobi algorithm is used in this function.
         */
-        void YARP_math_API SVDJacobi(const yarp::sig::Matrix &in, 
+        void YARP_math_API SVDJacobi(const yarp::sig::Matrix &in,
             yarp::sig::Matrix &U,
             yarp::sig::Vector &S,
             yarp::sig::Matrix &V);
 
         /**
         * Perform the moore-penrose pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param tol singular values less than tol are set to zero
         * @return pseudo-inverse of the matrix 'in'
         */
@@ -73,7 +73,7 @@ namespace yarp
 
         /**
         * Perform the moore-penrose pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param out pseudo-inverse of the matrix 'in'
         * @param tol singular values less than tol are set to zero
         */
@@ -81,7 +81,7 @@ namespace yarp
 
         /**
         * Perform the moore-penrose pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param sv vector containing the singular values of the input matrix
         * @param tol singular values less than tol are set to zero
         * @return pseudo-inverse of the matrix 'in'
@@ -90,7 +90,7 @@ namespace yarp
 
         /**
         * Perform the moore-penrose pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param out pseudo-inverse of the matrix 'in'
         * @param sv vector containing the singular values of the input matrix
         * @param tol singular values less than tol are set to zero
@@ -99,7 +99,7 @@ namespace yarp
 
         /**
         * Perform the damped pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param sv vector containing the singular values of the input matrix
         * @param damp damping factor
         */
@@ -107,14 +107,14 @@ namespace yarp
 
         /**
         * Perform the damped pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param damp damping factor
         */
         yarp::sig::Matrix YARP_math_API pinvDamped(const yarp::sig::Matrix &in, double damp);
 
         /**
         * Perform the damped pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param out damped pseudo-inverse of the matrix 'in'
         * @param damp damping factor
         */
@@ -122,7 +122,7 @@ namespace yarp
 
         /**
         * Perform the damped pseudo-inverse of a matrix (defined in SVD.h).
-        * @param in input matrix 
+        * @param in input matrix
         * @param out damped pseudo-inverse of the matrix 'in'
         * @param sv vector containing the singular values of the input matrix
         * @param damp damping factor
@@ -148,7 +148,7 @@ namespace yarp
         void YARP_math_API projectionMatrix(const yarp::sig::Matrix &A, yarp::sig::Matrix &out, double tol=0.0);
 
         /**
-        * Compute the nullspace projection matrix of A, that is defined as the difference between the 
+        * Compute the nullspace projection matrix of A, that is defined as the difference between the
         * identity matrix and the pseudoinverse of A times A: (I - pinv(A)*A) (defined in SVD.h).
         * Multiplying this null projection matrix times a vector projects the vector in the nullspace of A.
         * @param A input matrix
@@ -158,7 +158,7 @@ namespace yarp
         yarp::sig::Matrix YARP_math_API nullspaceProjection(const yarp::sig::Matrix &A, double tol=0.0);
 
         /**
-        * Compute the nullspace projection matrix of A, that is defined as the difference between the 
+        * Compute the nullspace projection matrix of A, that is defined as the difference between the
         * identity matrix and the pseudoinverse of A times A: (I - pinv(A)*A) (defined in SVD.h).
         * Multiplying this projection matrix times a vector projects the vector in the range of A.
         * @param A input matrix
